Creato componente payment_method
