+++ next-js/src/components/ingredient_chart.tsx
@@ -75,12 +75,6 @@ export function IngredientChart({ fetchedOrders, reservationId }: { fetchedOrder
     //TODO: riepilogo ordine? pagina apposta?
   }
 
-  const [selectedOption, setSelectedOption] = useState('AllaRomana');
-
-  const handleOptionChange = (option: string) => {
-    setSelectedOption(option);
-  };
-
   return (
     <>
       {Object.keys(orders).map((key: string) => (
@@ -108,49 +102,11 @@ export function IngredientChart({ fetchedOrders, reservationId }: { fetchedOrder
           </div>
         </div>
       ))}
-      <div className="w-full flex justify-center">
-        <div className="max-w-md w-full px-8 py-4 flex flex-col space-y-4 bg-white rounded-lg shadow-md">
-          <div className="text-lg font-semibold text-center">La tua parte:</div>
-
-          <div>
-            <label
-              htmlFor="AllaRomana"
-              className={`flex cursor-pointer items-center justify-between rounded-lg border border-gray-100 bg-white p-4 text-sm font-medium shadow-sm hover:border-gray-200 ${selectedOption === 'AllaRomana' ? 'border-orange-500 ring-1 ring-orange-500' : ''
-                }`}
-              onClick={() => handleOptionChange('AllaRomana')}
-            >
-              <p className="text-gray-700">Alla romana</p>
-              <input
-                type="radio"
-                className="sr-only"
-                id="AllaRomana"
-                checked={selectedOption === 'AllaRomana'}
-                onChange={() => handleOptionChange('AllaRomana')}
-              />
-            </label>
-          </div>
-
-          <div>
-            <label
-              htmlFor="Ognuno"
-              className="flex items-center justify-between rounded-lg bg-gray-100 p-4 text-sm font-medium shadow-sm"
-            >
-              <p className="text-gray-700">Ognun per sé (Coming Soon)</p>
-              <input
-                type="radio"
-                className="sr-only"
-                id="Ognuno"
-                disabled
-              />
-            </label>
-          </div>
 
           {/* Conferma */}
           <div className="flex justify-center mt-8">
             <button onClick={submit} className="bg-orange-500 hover:bg-orange-600 text-white px-4 py-2 rounded">Conferma</button>
           </div>
-        </div>
-      </div>
 
     </>
   );
